+++ Genesis/src/utopia/genesis/shape/shape2D/movement/MovementStatus2D.scala
@@ -1,5 +1,6 @@
-package utopia.genesis.shape.shape2D
+package utopia.genesis.shape.shape2D.movement
 
+import utopia.genesis.shape.shape2D.Vector2D
 import utopia.genesis.shape.template.MovementStatusLike
 
 /**