+++ Reflection/src/utopia/reflection/container/stack/template/scrolling/ScrollAreaLike2.scala
@@ -8,7 +8,8 @@ import utopia.genesis.handling.mutable.ActorHandler
 import utopia.genesis.shape.Axis._
 import utopia.genesis.shape._
 import utopia.genesis.shape.shape1D.LinearAcceleration
-import utopia.genesis.shape.shape2D.{Bounds, Point, Size, Velocity2D}
+import utopia.genesis.shape.shape2D.movement.Velocity2D
+import utopia.genesis.shape.shape2D.{Bounds, Point, Size}
 import utopia.genesis.shape.shape3D.Vector3D
 import utopia.genesis.shape.template.VectorLike
 import utopia.genesis.util.Drawer